feature: style and animation implemented
