+++ app/src/main/java/com/girlify/conversationApp/ui/CardText.kt
@@ -1,0 +1,22 @@
+package com.girlify.conversationApp.ui
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun CardText(text: String) {
+    Text(
+        text = text,
+        textAlign = TextAlign.Center,
+        fontWeight = FontWeight.Bold,
+        fontSize = 42.sp,
+        lineHeight = 48.sp,
+        modifier = Modifier.padding(horizontal = 16.dp)
+    )
+}
