+++ app/src/main/java/com/girlify/conversationApp/ui/theme/Color.kt
@@ -6,6 +6,6 @@ val Purple80 = Color(0xFFD0BCFF)
 val PurpleGrey80 = Color(0xFFCCC2DC)
 val Pink80 = Color(0xFFEFB8C8)
 
-val FederalBlue = Color(0xFF30347F)
-val PurpleGrey40 = Color(0xFF625b71)
+val Blue800 = Color(0xFF30347F)
+val Pink500 = Color(0xFFC1007C)
 val Pink40 = Color(0xFF7D5260)
